+++ .github/workflows/dart.yaml
@@ -46,7 +46,7 @@ jobs:
 
           find lib \
           '!' -name '*_web.dart' \
-          -name '*dart' | 
+          -name '*dart' |
           cut -c4- |
           awk -v package=$1 '{printf "import '\''package:flutter_code_editor%s%s'\'';\n", package, $1}' >> $file
           echo "void main(){}" >> $file
