+++ lib/src/code_field/code_field.dart
@@ -15,7 +15,7 @@ import 'actions/indent.dart';
 import 'actions/outdent.dart';
 import 'code_controller.dart';
 import 'default_styles.dart';
-import 'disable_spell_check.dart';
+import 'disable_spell_check/disable_spell_check.dart';
 
 final _shortcuts = <ShortcutActivator, Intent>{
   // Copy
